requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
filelock==3.16.1
fsspec==2024.12.0
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.1.1
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.1
packaging==24.2
pgvector==0.3.6
phidata==2.7.6
pip==24.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.1
pygments==2.19.1
pypdf==5.1.0
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.36
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing-extensions==4.12.2
urllib3==2.3.0
uv==0.5.15
